requirements.txt
requests==2.32.3
PyYAML==6.0.2
mysql-connector-python==9.1.0
python-dotenv==1.0.1
schedule==1.2.2